backend/jetson-requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
audioread==2.1.8
baidu-aip==2.2.18.0
beautifulsoup4==4.9.3
bidict==0.21.2
blis==0.7.4
bs4==0.0.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.2
cffi==1.14.5
cfgv==3.2.0
chardet==3.0.4
CJKwrap==2.2
click==7.1.2
ConfigArgParse==1.4
configparser==5.0.1
coverage==5.4
ctranslate2 # must be installed from source
cymem==2.0.3
decorator==4.4.2
distlib==0.3.1
dnspython==1.16.0
docker-pycreds==0.4.0
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
gevent==20.12.1
gevent-websocket==0.10.1
geventhttpclient==1.4.4
gitdb==4.0.5
GitPython==3.1.13
google-api-core==1.31.0
google-auth==1.32.1
google-cloud-core==1.4.1
google-cloud-speech==1.3.2
google-cloud-translate==3.0.1
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.29.0
gunicorn==20.0.4
huggingface-hub==0.0.13
HyperPyYAML==1.0.0
identify==2.2.4
idna==2.9
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jiwer==2.2.0
joblib==0.15.1
libcst==0.3.10
locust==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mosestokenizer==1.1.0
msgpack==1.0.2
multidict==4.7.6
murmurhash==1.0.2
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.18.5
opencv-python==4.4.0.42
openfile==0.0.7
packaging==21.0
pathtools==0.1.2
plac==1.1.3
pluggy==0.13.1
portalocker==2.0.0
pre-commit==2.12.1
preshed==3.0.2
promise==2.3
proto-plus==1.7.1
protobuf==3.12.2
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.13.2
python-Levenshtein==0.12.0
python-socketio==4.6.0
python-speech-features==0.6
pytz==2020.1
PyYAML==5.3.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.24.0
rsa==4.6
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sacrebleu==1.4.14
sacremoses==0.0.43
scikit-learn==0.23.1
scipy
sentencepiece==0.1.96
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
SoundFile==0.10.3.post1
soupsieve==2.1
spacy==2.3.7
speechbrain==0.5.9
srsly==1.0.2
subprocess32==3.5.4
termcolor==1.1.0
thinc==7.4.5
threadpoolctl==2.1.0
toml==0.10.2
toolwrapper==2.1.0
torch==1.9.0
torchaudio==0.9.0 # set `--extra-index-url https://download.pytorch.org/whl/test/cpu` for the arm version when installing with pip. see https://github.com/pytorch/audio/issues/1594
tqdm==4.50.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uctools==1.3.0
urllib3==1.25.9
virtualenv==20.4.4
wandb==0.10.18
wasabi==0.8.0
websocket-client==0.58.0
Werkzeug==1.0.1
yarl==1.5.1
zipp==3.3.1
zope.event==4.5.0
zope.interface==5.2.0
yake
zhon
